move around vidstack styles
