+++ packages/gamut/src/Video/lib/VidstackPlayer/index.tsx
@@ -1,9 +1,8 @@
-/* Third-party Vidstack default theme styles */
+/* Vidstack default theme (vendored via @import) + gamut overrides */
 /* eslint-disable gamut/no-css-standalone */
-import '../vidstack-vendor.css';
+import './vidstack-styles.css';
 
 import { styledOptions } from '@codecademy/gamut-styles';
-import { Global } from '@emotion/react';
 import styled, { CSSObject } from '@emotion/styled';
 import {
   isYouTubeProvider,
@@ -18,12 +17,11 @@ import {
 } from '@vidstack/react';
 import React, { useRef } from 'react';
 
-import { Box } from '../../Box';
-import { VideoProps } from '..';
-import { vdsBaseThemeStyles } from '../styles';
-import { keyboardShortcuts } from './utils/constants';
-import { vdsVariables } from './utils/variables';
-import { VideoLayout } from './VideoLayout';
+import { Box } from '../../../Box';
+import { VideoProps } from '../..';
+import { keyboardShortcuts } from '../utils/constants';
+import { VideoLayout } from '../VideoLayout';
+import { vdsVariables } from './variables';
 
 const VariableProvider = styled(Box, styledOptions(['variables']))<{
   variables?: CSSObject;
@@ -83,7 +81,6 @@ export const VidstackPlayer: React.FC<VidstackPlayerProps> = ({
       style={{ width, height }}
       variables={vdsVariables}
     >
-      <Global styles={vdsBaseThemeStyles} />
       <MediaPlayer
         autoPlay={autoplay}
         controls={showDefaultProviderControls}
