--- packages/gamut/src/Video/vidstack-vendor.css
@@ -1,3 +1,0 @@
-@import "@vidstack/react/player/styles/default/theme.css";
-@import "@vidstack/react/player/styles/default/layouts/video.css";
-@import "@vidstack/react/player/styles/default/layouts/audio.css";
